Separate actions for users to own file
