+++ client/src/index.js
@@ -7,7 +7,7 @@ import { createStore, applyMiddleware } from "redux";
 import reduxThunk from "redux-thunk";
 import jwt_decode from "jwt-decode";
 import setAuthToken from "./js/utils/setAuthToken";
-import { setCurrentUser, logoutUser } from "./js/actions/index";
+import { setCurrentUser, logoutUser } from "./js/actions/users";
 import { Provider } from "react-redux";
 import reducers from "./js/reducers";
 import { BrowserRouter } from "react-router-dom";
