+++ client/src/js/components/Dashboard/Dashboard.test.js
@@ -4,10 +4,10 @@ import { MemoryRouter } from 'react-router-dom';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import { Dashboard } from './Dashboard';
-import { logoutUser } from "../../actions/index";
+import { logoutUser } from "../../actions/users";
 
 Enzyme.configure({ adapter: new Adapter() });
-jest.mock('../../actions/index');
+jest.mock('../../actions/users');
 
 describe('Dashboard', () => {
 	const params = {
