+++ client/src/js/components/Register/Register.test.js
@@ -4,10 +4,10 @@ import { MemoryRouter } from 'react-router-dom';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import { Register } from './Register';
-import { registerUser } from "../../actions/index";
+import { registerUser } from "../../actions/users";
 
 Enzyme.configure({ adapter: new Adapter() });
-jest.mock('../../actions/index');
+jest.mock('../../actions/users');
 
 describe('Register', () => {
 	const params = {
